packages/experiment/pyproject.toml
[project]
name = "experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "agent-framework>=1.0.0b251120",
    # needed for logging in AzureML
    "agent-framework-azure-ai>=1.0.0b251120",
    "azure-ai-projects>=1.0.0",
    "azureml-mlflow==1.60.*",
    "jinja2>=3.1.6",
    "mlflow-skinny==2.21.*",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pytz==2025.2",
]

[project.scripts]
experiment = "experiment.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
